backend/requirements.txt
Django>=4.2
djangorestframework>=3.14.0
python-decouple>=3.8
google-api-python-client>=2.107.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
django-cors-headers
